ask: add attachment capability
When ask sees a line in a file starting with <<, it will assume the rest of the line is a filename.
Ask will attach the contents of filename to the message, prepending it to the rest of the question file content.
Note that the attachment lines themselves are not included in messages to the AI.